# backend/public/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "To-Do List API",
		"description": "REST API для управления задачами. Laravel + Sanctum Bearer Token.",
		"version": "1.1.0"
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Local development"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "sanctum"
			}
		},
		"schemas": {
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"name": {
						"type": "string",
						"example": "Admin User"
					},
					"email": {
						"type": "string",
						"example": "admin@example.com"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin"
						],
						"example": "admin"
					},
					"is_admin": {
						"type": "boolean",
						"example": true
					}
				}
			},
...
			"ProfileUpdateRequest": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"example": "John Doe",
						"description": "Новое имя (опционально)"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "new@example.com",
						"description": "Новый email (опционально, проверка уникальности)"
					},
					"password": {
						"type": "string",
						"format": "password",
						"minLength": 6,
						"example": "newpass",
						"description": "Новый пароль (опционально)"
					}
				}
			},
			"ProfileUpdateResponse": {
				"type": "object",
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"message": {
						"type": "string",
						"example": "Профиль обновлён."
					}
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string"
					},
					"errors": {
						"type": "object",
						"additionalProperties": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				}
			}
		}
	},
	"paths": {
...
		"/api/user": {
			"get": {
				"tags": [
					"Auth"
				],
				"summary": "Текущий пользователь",
				"description": "Возвращает данные авторизованного пользователя.",
				"operationId": "user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Данные пользователя",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Не авторизован",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"Auth"
				],
				"summary": "Обновление профиля",
				"description": "Обновление имени, email и/или пароля текущего пользователя. Email проверяется на уникальность.",
				"operationId": "updateUser",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProfileUpdateRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Профиль обновлён",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileUpdateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Не авторизован",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Ошибка валидации",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	}
}